add nationalize charset and modify extraction upload algorithm

+++ src/main/java/com/example/mssqll/dto/response/ExtractionResponseDto.java
@@ -16,7 +16,7 @@ public class ExtractionResponseDto {
 
     private LocalDate date;
 
-    private int totalAmount;
+    private Double totalAmount;
 
     private String purpose;
